# src/main/resources/static/api-docs/api-doc.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/v1/medicos/{id}": {
      "get": {
        "tags": [
          "Médico"
        ],
        "summary": "Busca um médico pelo ID",
        "operationId": "buscarMedico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Médico encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medico"
                }
              }
            }
          },
          "404": {
            "description": "Médico não encontrado"
          }
        }
      },
      "put": {
        "tags": [
          "Médico"
        ],
        "summary": "Atualiza um médico",
        "operationId": "atualizarMedico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Médico inválido"
          },
          "200": {
            "description": "Médico Atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medico"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Médico"
        ],
        "summary": "Deleta um médico",
        "operationId": "deletarMedico",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Médico deletado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medico"
                }
              }
            }
          },
          "404": {
            "description": "Médico não encontrado"
          }
        }
      }
    },
...
    "/v1/medicos": {
      "post": {
        "tags": [
          "Médico"
        ],
        "summary": "Cria um médico",
        "operationId": "criarMedico",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicoDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "description": "Médico inválido"
          },
          "201": {
            "description": "Médico Criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medico"
                }
              }
            }
          }
        }
      }
    },
    "/v1/medicos/{idMedico}/especialidades/{idEspecialidade}": {
      "post": {
        "tags": [
          "Médico"
        ],
        "summary": "Vincula uma especialidade a um médico",
        "operationId": "vincularEspecialidade",
        "parameters": [
          {
            "name": "idMedico",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "idEspecialidade",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Especialidade vinculada ao médico",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medico"
                }
              }
            }
          },
          "400": {
            "description": "Requisição inválida"
          },
          "404": {
            "description": "Médico ou especialidade não encontrada"
          }
        }
      },
      "delete": {
        "tags": [
          "Médico"
        ],
        "summary": "Desvincula uma especialidade de um médico",
        "operationId": "desvincularEspecialidade",
        "parameters": [
          {
            "name": "idMedico",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "idEspecialidade",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Requisição inválida"
          },
          "404": {
            "description": "Médico ou especialidade não encontrada"
          },
          "200": {
            "description": "Especialidade desvinculada do médico",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medico"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/medicos/especialidade/{idEspecialidade}": {
      "get": {
        "tags": [
          "Médico"
        ],
        "summary": "Busca médicos por especialidade",
        "operationId": "buscarMedicosPorEspecialidade",
        "parameters": [
          {
            "name": "idEspecialidade",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Médicos encontrados",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "204": {
            "description": "Nenhum médico encontrado"
          }
        }
      }
    },
    "/v1/medicos/email/{email}": {
      "get": {
        "tags": [
          "Médico"
        ],
        "summary": "Busca um médico pelo email",
        "operationId": "buscarMedicoPorEmail",
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Médico encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medico"
                }
              }
            }
          },
          "404": {
            "description": "Médico não encontrado"
          }
        }
      }
    },
    "/v1/medicos/crm/{crm}": {
      "get": {
        "tags": [
          "Médico"
        ],
        "summary": "Busca um médico pelo CRM",
        "operationId": "buscarMedicoPorCrm",
        "parameters": [
          {
            "name": "crm",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Médico encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Medico"
                }
              }
            }
          },
          "404": {
            "description": "Médico não encontrado"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EspecialidadeDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codigo": {
            "type": "integer",
            "format": "int64"
          },
          "descricao": {
            "type": "string"
          }
        }
      },
      "MedicoDTO": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "crm": {
            "type": "string"
          },
          "especialidades": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EspecialidadeDTO"
            }
          }
        }
      },
      "Especialidade": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "codigo": {
            "type": "integer",
            "format": "int64"
          },
          "descricao": {
            "type": "string"
          },
          "dataCriacao": {
            "type": "string",
            "format": "date-time"
          },
          "dataAtualizacao": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Medico": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "nome": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "dataCriacao": {
            "type": "string",
            "format": "date-time"
          },
          "dataAtualizacao": {
            "type": "string",
            "format": "date-time"
          },
          "crm": {
            "type": "string"
          },
          "especialidades": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Especialidade"
            }
          }
        }
      },
...
    }
  }
}